Refactor module imports
