+++ index.js
@@ -5,16 +5,6 @@ const https = require('https');
 const url = require('url');
 
 /* Require own modules */
-const fbAuth = require('./services/server_authentication.js');
-const fbGraph = require('./services/facebook_query.js');
-const mongoActions = require('./services/mongo_client.js');
-const mongoConnect = require('./services/mongo_client.js').dbConnect;
-const mongoInsert = mongoActions.insertDocuments;
-const getGroup = fbGraph.getGroup;
-const getFeed = fbGraph.getGroupFeed;
-const getReactions = fbGraph.getReactions;
-const getComments = fbGraph.getComments;
-
 const objectController = require('./controllers/objectController.js');
 const feedController = require('./controllers/feedController.js');
 const reactionsController = require('./controllers/reactionsController.js');
@@ -25,91 +15,6 @@ const commentsController = require('./controllers/commentsController.js');
 const hostname = '127.0.0.1';
 const port = '8129';
 
-var feedRes = '';
-/**
- * Returns a Promise that contains an array of posts, with
- * their respective ids, story and reactions
- *
- * @param {Array} promises - Array of promises
- * @param {Array} ids - Array of ids
- * @param {Array} stories - Array of stories
- *
- * @returns {Promise} Promise that resolves with an array of Posts and metadata
- *
- */
-const _promiseAccumulator = function(promises, ids, stories, dates) {
-  let returnObj = [];
-  var ready = Promise.resolve(null);
-
-  promises.forEach(function _promiseIteration(promise, index) {
-    ready = ready.then(function _readyCallback() {
-      return promise;
-    })
-    .then(value => {
-      //console.log(value)
-      var individualObj = {};
-      individualObj['post_id'] = `${ids[index]}`;
-      individualObj['date_posted'] = `${dates[index]}`;
-      individualObj['post_message'] = `${stories[index]}`;
-      individualObj['post_reactions'] = JSON.parse(value).data;
-      returnObj.push(individualObj);
-    })
-  })
-  return ready.then(() => { return returnObj; });
-}
-
-/**
- * Returns a Promise that contains an array of posts, with
- * their respective ids, story and reactions
- *
- * @param {Array} data - Array of post objects
- *
- * @returns {Object} Object containing Promises, Ids, and Stories
- *
- */
-const _mapDataToReaction = function(data) {
-  let promiseArr = []
-  let idsArr = []
-  let storiesArr = []
-  let dateArray = []
-  data.map(function(curr, index, arr) {
-    promiseArr.push(getReactions(curr.id))
-    idsArr.push(curr.id)
-    storiesArr.push((curr.story) ? curr.story : curr.message );
-    dateArray.push(curr.created_time)
-  })
-  return {
-    promiseArr: promiseArr,
-    idsArr: idsArr,
-    storiesArr: storiesArr,
-    dates: dateArray
-  }
-}
-
-/**
- *
- *
- */
-
-const _mapDataToComments = function(postArray) {
-  let getCommentsPromises = [];
-  postArray.map(function(curr, index, arr) {
-    getCommentsPromises.push(getComments(curr.post_id, 'summary=1&filter=stream'))
-  });
-  return Promise.all(getCommentsPromises);
-}
-
-const _handleCommentsArray = function(curr, ndx, arr) {
-  return JSON.parse(curr).summary.total_count;
-}
-
-const _mergeTwoArrays = function(arr1, arr2) {
-  let retArr = arr1.map(function(curr, ndx, arr) {
-    curr['post_comment_total'] = arr2[ndx];
-    return curr;
-  })
-  return retArr;
-}
 
 const server = http.createServer(function serverCallback (req, res) {
   var uri = url.parse(req.url);
@@ -124,85 +29,19 @@ const server = http.createServer(function serverCallback (req, res) {
       });
   }
   if(path.indexOf('object') > 0) {
-    var groupResponse = '';
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    var groupId = path.split('/')[2]
-    getGroup(groupId)
-    .then(response => {
-      groupResponse = response.toString()
-      res.end(groupResponse);
-    })
-    .catch(err => {
-      console.log(err)
-    });
+    objectController(res, path, uri);
   }
   if(path.indexOf('feed') > 0) {
-    var groupResponse = '';
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    var groupId = path.split('/')[2]
-    getFeed(groupId, uri.query)
-    .then(response => {
-      groupResponse = response.toString()
-      res.end(groupResponse);
-    })
-    .catch(err => {
-      console.log(err)
-    });
+    feedController(res, path, uri);
   }
   if(path.indexOf('reactions') > 0) {
-    var reactionResponse = '';
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    var groupId = path.split('/')[2];
-    getReactions(groupId, uri.query)
-    .then(response => {
-      reactionResponse = response.toString();
-      res.end(reactionResponse);
-    })
-    .catch(err => {
-      console.log(err);
-    })
+    feedController(res, path, uri);
   }
   if(path.indexOf('meta') > 0 ) {
-
-    let feedResponse = '';
-    let feedData = '';
-    let groupId = path.split('/')[2];
-
-    getFeed(groupId, uri.query)
-    .then(response => {
-      feedResponse = response.toString();
-
-      feedData = JSON.parse(feedResponse).data
-      promiseData = _mapDataToReaction(feedData)
-
-      _promiseAccumulator(
-        promiseData.promiseArr,
-        promiseData.idsArr,
-        promiseData.storiesArr,
-        promiseData.dates
-      )
-      .then(value => {
-        /* value == [{posts},{posts}] */
-        _mapDataToComments(value)
-        .then(withComments => {
-          let commentSummary = withComments.map(_handleCommentsArray)
-          let posts = _mergeTwoArrays(value, commentSummary)
-          mongoInsert(posts)
-          .then(r => { console.log(r) })
-          .catch(err => { console.log(err) })
-          res.end(JSON.stringify(posts))
-        })
-      })
-    })
-    .catch(err => {
-      console.log(err)
-    });
+    metaController(res, path, uri)
   }
   if(path.indexOf('comments') > 0) {
-    commentsController(res, path, uri)
+    commentsController(res, path, uri);
   }
 });
